third_party/blink/web_tests/fast/forms/focus-on-control-with-zero-size-expected.txt

Check if a control won't loose focus when it becomes 0-size.
PASS document.activeElement is input1
PASS input1.value is "01"
PASS document.activeElement is input1


Check if 0-size control can get focus.
PASS document.activeElement is input1
PASS successfullyParsed is true

TEST COMPLETE
